+++ tests/testthat/test-class_parquet.R
@@ -138,7 +138,7 @@ tar_test("store_path_from_record()", {
 })
 
 tar_test("parquet and NULL", {
-  skip_on_cran()
+  skip_cran()
   tar_script(
     list(
       tar_target(x, NULL, format = "parquet", memory = "persistent"),